Move url_config to conf.service_url
