+++ tests/utils_tests/test_url_config.py
@@ -4,7 +4,7 @@
 import unittest
 
 from django.db import connection
-from django.utils.url_config import configure_cache, configure_db, parse_url
+from django.conf.service_url import configure_cache, configure_db, parse_url
 
 GENERIC_TESTS = [
     (
